Refactor nemoguardrails.eval into nemoguardrails.evaluate.
