+++ docs/evaluation/README.md
@@ -45,7 +45,7 @@ For the initial evaluation experiments for dialog rails, we have used two datase
 
 The datasets were transformed into a NeMo Guardrails app by defining canonical forms for each intent, specific dialogue flows, and even bot messages (for the _chit-chat_ dataset alone).
 The two datasets have a large number of user intents, thus dialog rails. One of them is very generic and has higher-grained intents (_chit-chat_), while the _banking_ dataset is domain-specific and more fine-grained.
-More details about running the dialog rails evaluation experiments and the evaluation datasets are available [here](./../../nemoguardrails/eval/data/topical/README.md).
+More details about running the dialog rails evaluation experiments and the evaluation datasets are available [here](../../nemoguardrails/evaluate/data/topical/README.md).
 
 Preliminary evaluation results follow next. In all experiments, we have chosen to have a balanced test set with at most 3 samples per intent.
 For both datasets, we have assessed the performance for various LLMs and also for the number of samples (`k = all, 3, 1`) per intent that are indexed in the vector database.
@@ -163,7 +163,7 @@ Here is a list of arguments that you can use to configure the fact-checking rail
 - `output-dir`: The directory to save the output to. The default is `eval_outputs/factchecking`.
 - `write-outputs`: Whether to write the outputs to a file or not. The default is `True`.
 
-More details on how to set up the data in the right format and run the evaluation on your own dataset can be found [here](./../../nemoguardrails/eval/data/factchecking/README.md).
+More details on how to set up the data in the right format and run the evaluation on your own dataset can be found [here](../../nemoguardrails/evaluate/data/factchecking/README.md).
 
 #### Evaluation Results
 
@@ -171,7 +171,7 @@ Evaluation Date - Nov 23, 2023 (Mar 7 2024 for `gemini-1.0-pro`).
 
 We evaluate the performance of the fact-checking rail on the [MSMARCO](https://huggingface.co/datasets/ms_marco) dataset using the Self-Check and the AlignScore approaches. To build the dataset, we randomly sample 100 (question, correct answer, evidence) triples, and then, for each triple, build a non-factual or incorrect answer to yield 100 (question, incorrect answer, evidence) triples.
 
-We breakdown the performance into positive entailment accuracy and negative entailment accuracy. Positive entailment accuracy is the accuracy of the model in correctly identifying answers that are grounded in the evidence passage. Negative entailment accuracy is the accuracy of the model in correctly identifying answers that are **not** supported in the evidence. Details on how to create synthetic negative examples can be found [here](./../../nemoguardrails/eval/data/factchecking/README.md)
+We breakdown the performance into positive entailment accuracy and negative entailment accuracy. Positive entailment accuracy is the accuracy of the model in correctly identifying answers that are grounded in the evidence passage. Negative entailment accuracy is the accuracy of the model in correctly identifying answers that are **not** supported in the evidence. Details on how to create synthetic negative examples can be found [here](../../nemoguardrails/evaluate/data/factchecking/README.md)
 
 | Model                  | Positive Entailment Accuracy | Negative Entailment Accuracy | Overall Accuracy | Average Time Per Checked Fact (ms) |
 |------------------------|------------------------------|------------------------------|------------------|------------------------------------|
@@ -224,7 +224,7 @@ To evaluate the output moderation rail only, use the following command:
 
 ```nemoguardrails evaluate moderation --check-input False --config=path/to/guardrails/config```
 
-More details on how to set up the data in the right format and run the evaluation on your own dataset can be found [here](./../../nemoguardrails/eval/data/moderation/README.md).
+More details on how to set up the data in the right format and run the evaluation on your own dataset can be found [here](../../nemoguardrails/evaluate/data/moderation/README.md).
 
 #### Evaluation Results
 
@@ -315,7 +315,7 @@ To evaluate the hallucination rail on your own dataset, you can follow the creat
 
 #### Evaluation Results
 
-To evaluate the hallucination rail, we manually curate a set of [questions](./../../nemoguardrails/eval/data/hallucination/sample.txt) which mainly consists of questions with a false premise, i.e., questions that cannot have a correct answer.
+To evaluate the hallucination rail, we manually curate a set of [questions](../../nemoguardrails/evaluate/data/hallucination/sample.txt) which mainly consists of questions with a false premise, i.e., questions that cannot have a correct answer.
 
 For example, the question "What is the capital of the moon?" has a false premise since the moon does not have a capital. Since the question is stated in a way that implies that the moon has a capital, the model might be tempted to make up a fact and answer the question.
