+++ examples/configs/llm/README.md
@@ -4,6 +4,6 @@ The examples in this folder showcase how various LLM providers should be configu
 
 The guardrails configurations are very simple and only meant to show that the LLM is configured properly. These configurations **ARE NOT** meant to be used for production deployments.
 
-Some LLM providers showcased in the folder are included in the topical rails evaluation and achieve good results for the proposed datasets. See the results [here](../../../nemoguardrails/eval/README.md).
+Some LLM providers showcased in the folder are included in the topical rails evaluation and achieve good results for the proposed datasets. See the results [here](../../../nemoguardrails/evaluate/README.md).
 
 The team encourages and welcomes contributions from the community. LLM-specific configurations like the ones shown here are a great way to get involved. If you see opportunities to improve the way NeMo Guardrails works with common LLMs, please consider submitting a PR.
