+++ examples/configs/llm/hf_pipeline_falcon/README.md
@@ -2,7 +2,7 @@
 
 This configuration uses the HuggingFace Pipeline LLM with a [tiiuae/falcon-7b-instruct](https://huggingface.co/tiiuae/falcon-7b-instruct) model.
 
-The `tiiuae/falcon-7b-instruct` LLM model has been tested on the topical rails evaluation sets, results are available [here](../../../../nemoguardrails/eval/README.md).
+The `tiiuae/falcon-7b-instruct` LLM model has been tested on the topical rails evaluation sets, results are available [here](../../../../nemoguardrails/evaluate/README.md).
 
 In this folder, the guardrails application is very basic, but anyone can change it with any other more complex configuration.
