+++ examples/configs/llm/hf_pipeline_mosaic/README.md
@@ -2,7 +2,7 @@
 
 This configuration uses the HuggingFace Pipeline LLM with the [mpt-7b-instruct](https://huggingface.co/mosaicml/mpt-7b-instruct) model.
 
-The `mpt-7b-instruct` LLM model has been tested on the topical rails evaluation sets, results are available [here](../../../../nemoguardrails/eval/README.md).
+The `mpt-7b-instruct` LLM model has been tested on the topical rails evaluation sets, results are available [here](../../../../nemoguardrails/evaluate/README.md).
 
 In this folder, the guardrails application is very basic, but anyone can change it with any other more complex configuration.
