+++ nemoguardrails/evaluate/evaluate_hallucination.py
@@ -22,7 +22,7 @@ import tqdm
 import typer
 
 from nemoguardrails import LLMRails
-from nemoguardrails.eval.utils import load_dataset
+from nemoguardrails.evaluate.utils import load_dataset
 from nemoguardrails.llm.params import llm_params
 from nemoguardrails.llm.prompts import Task
 from nemoguardrails.llm.taskmanager import LLMTaskManager
